# apps/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organisation/teams": {
      "post": {
        "operationId": "TeamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Create a team",
        "tags": [
          "Organisation / Teams"
        ]
      },
      "get": {
        "operationId": "TeamsController_search",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by title or description",
            "schema": {
              "minLength": 1,
              "maxLength": 255,
              "example": "Engineering Team",
              "type": "string"
            }
          },
          {
            "name": "headId",
            "required": false,
            "in": "query",
            "description": "Id of team leader",
            "schema": {
              "example": 12,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sorting field",
            "schema": {
              "default": "id",
              "example": "title",
              "type": "string",
              "enum": [
                "id",
                "title",
                "description",
                "headId",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "Sorting direction",
            "schema": {
              "default": "asc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Search teams",
        "tags": [
          "Organisation / Teams"
        ]
      }
    },
    "/organisation/teams/{id}": {
      "get": {
        "operationId": "TeamsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Get a team by id",
        "tags": [
          "Organisation / Teams"
        ]
      },
      "patch": {
        "operationId": "TeamsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Update a team",
        "tags": [
          "Organisation / Teams"
        ]
      },
      "delete": {
        "operationId": "TeamsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Delete a team",
        "tags": [
          "Organisation / Teams"
        ]
      }
    },
    "/organisation/teams/{id}/members": {
      "post": {
        "operationId": "TeamsController_addMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTeamMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Add a member to a team",
        "tags": [
          "Organisation / Teams"
        ]
      },
      "get": {
        "operationId": "TeamsController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get the team members",
        "tags": [
          "Organisation / Teams"
        ]
      }
    },
    "/organisation/teams/{id}/members/{userId}": {
      "delete": {
        "operationId": "TeamsController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Delete a member from a team",
        "tags": [
          "Organisation / Teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "Intra API",
    "description": "API documentation for Intra",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "example": "Valerii"
          },
          "secondName": {
            "type": "object",
            "example": "Velychko",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "example": "Valeriiovych"
          },
          "fullName": {
            "type": "object",
            "example": "Valerii Valeriiovych Velychko",
            "nullable": true
          },
          "email": {
            "type": "string",
            "example": "valerii.velychko@example.com"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ],
            "example": "active"
          },
          "positionId": {
            "type": "number",
            "example": 1,
            "nullable": true
          },
          "teamId": {
            "type": "number",
            "example": 10,
            "nullable": true
          },
          "managerId": {
            "type": "number",
            "example": 2,
            "nullable": true
          },
          "roles": {
            "type": "array",
            "example": [
              "manager"
            ],
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "hr",
                "manager",
                "employee"
              ]
            }
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-02T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "firstName",
          "secondName",
          "lastName",
          "fullName",
          "email",
          "status",
          "positionId",
          "teamId",
          "managerId",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateTeamDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Team title",
            "example": "Engineering Team",
            "maxLength": 255,
            "minLength": 1
          },
          "description": {
            "type": "object",
            "description": "Team description",
            "example": "Responsible for product development",
            "nullable": true,
            "maxLength": 1000,
            "minLength": 1
          },
          "headId": {
            "type": "object",
            "description": "Id of team leader",
            "example": 12,
            "nullable": true
          }
        },
        "required": [
          "title"
        ]
      },
      "TeamResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Engineering Team"
          },
          "description": {
            "type": "object",
            "example": "Responsible for product development",
            "nullable": true
          },
          "headId": {
            "type": "object",
            "example": 10,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-02T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "headId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTeamDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Team title",
            "example": "Engineering Team",
            "maxLength": 255,
            "minLength": 1
          },
          "description": {
            "type": "object",
            "description": "Team description",
            "example": "Responsible for product development",
            "nullable": true,
            "maxLength": 1000,
            "minLength": 1
          },
          "headId": {
            "type": "object",
            "description": "Id of team leader",
            "example": 12,
            "nullable": true
          }
        }
      },
      "AddTeamMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "User id",
            "example": 15
          },
          "isPrimary": {
            "type": "boolean",
            "description": "Is primary membership",
            "example": true
          }
        },
        "required": [
          "userId"
        ]
      },
      "TeamMemberResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 101
          },
          "teamId": {
            "type": "number",
            "example": 3
          },
          "userId": {
            "type": "number",
            "example": 42
          },
          "isPrimary": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-02T10:00:00.000Z"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "id",
          "teamId",
          "userId",
          "isPrimary",
          "createdAt"
        ]
      },
...
    }
  }
}